exps/requirements.txt




Pillow==10.1.0
# torch==2.5.1+cu121
# torch
torchaudio==2.5.1
torchvision==0.20
transformers==4.44.2
librosa==0.9.0
soundfile==0.12.1
vector-quantize-pytorch==1.18.5
vocos==0.1.0
decord
moviepy
